+++ src/annotation/multi_inject.ts
@@ -1,20 +1,6 @@
 import * as METADATA_KEY from "../constants/metadata_keys";
-import { interfaces } from "../interfaces/interfaces";
-import { Metadata } from "../planning/metadata";
-import { tagParameter, tagProperty } from "./decorator_utils";
+import { injectBase } from "./inject_base";
 
-function multiInject(serviceIdentifier: interfaces.ServiceIdentifier<any>) {
-  return function(target: any, targetKey: string, index?: number) {
-
-    const metadata = new Metadata(METADATA_KEY.MULTI_INJECT_TAG, serviceIdentifier);
-
-    if (typeof index === "number") {
-      tagParameter(target, targetKey, index, metadata);
-    } else {
-      tagProperty(target, targetKey, metadata);
-    }
-
-  };
-}
+const multiInject = injectBase(METADATA_KEY.MULTI_INJECT_TAG);
 
 export { multiInject };