+++ src/constants/metadata_keys.ts
@@ -31,6 +31,9 @@ export const DESIGN_PARAM_TYPES = "design:paramtypes";
 // used to identify postConstruct functions
 export const POST_CONSTRUCT = "post_construct";
 
+// used to identify preDestroy functions
+export const PRE_DESTROY = "pre_destroy";
+
 function getNonCustomTagKeys(): string[] {
     return [
         INJECT_TAG,