+++ src/inversify.ts
@@ -3,15 +3,18 @@ export const METADATA_KEY = keys;
 export { Container } from "./container/container";
 export { BindingScopeEnum, BindingTypeEnum, TargetTypeEnum } from "./constants/literal_types";
 export { AsyncContainerModule, ContainerModule } from "./container/container_module";
+export { createTaggedDecorator } from "./annotation/decorator_utils"
 export { injectable } from "./annotation/injectable";
 export { tagged } from "./annotation/tagged";
 export { named } from "./annotation/named";
-export { inject, LazyServiceIdentifer } from "./annotation/inject";
+export { inject } from "./annotation/inject";
+export { LazyServiceIdentifer } from "./annotation/lazy_service_identifier"
 export { optional } from "./annotation/optional";
 export { unmanaged } from "./annotation/unmanaged";
 export { multiInject } from "./annotation/multi_inject";
 export { targetName } from "./annotation/target_name";
 export { postConstruct } from "./annotation/post_construct";
+export { preDestroy } from "./annotation/pre_destroy";
 export { MetadataReader } from "./planning/metadata_reader";
 export { id } from "./utils/id";
 export { interfaces } from "./interfaces/interfaces";