+++ src/utils/factory_type.ts
@@ -1,0 +1,5 @@
+export enum FactoryType {
+    DynamicValue = "toDynamicValue",
+    Factory = "toFactory",
+    Provider = "toProvider",
+}