+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - only inject decorator can be applied to setters #1342
 - Container.resolve should resolve in that container #1338
+
+## [Unreleased]
+-Feat add API method for check dependency only in current container
+
 ## [5.1.1] - 2021-04-25
 -Fix pre-publish for build artifacts