+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Upgrade information for v4.x to v5.x
 
+### Changed
+- Update BindingToSyntax with `.toAutoNamedFactory()`.
+
 ### Fixed
 - Fix `Target.isTagged()` to exclude `optional` from tag injections #1190.
 - Update `toConstructor`, `toFactory`, `toFunction`, `toAutoFactory`, `toProvider` and `toConstantValue` to have singleton scope #1297.